Allow import/export spellbook
