+++ src/components/SpellBook.tsx
@@ -2,6 +2,7 @@ import "./SpellBook.css";
 
 import { APP_KEY, ASSET_LOCATION } from "../config";
 import { FaCaretDown, FaCaretUp, FaCirclePlus, FaDownload, FaFloppyDisk, FaPencil, FaTrash, FaUpload } from "react-icons/fa6";
+import { downloadFileFromString, loadJSONFile } from "../utils";
 import { getAllSpellNames, getSpell, spellIDs } from "../effects/spells";
 import { setSelectedSpell, toolID } from "../effectsTool";
 import { useCallback, useEffect, useRef, useState } from "react";
@@ -25,6 +26,7 @@ export default function SpellBook() {
     const [editing, setEditing] = useState(false);
     const [isGM, setIsGM] = useState(false);
     const mainDiv = useRef<HTMLDivElement>(null);
+    const fileInputRef = useRef<HTMLInputElement>(null);
 
     const openModal = (modalName: ModalType) => {
         setIsModalClosing(false);
@@ -134,6 +136,7 @@ export default function SpellBook() {
 
     return <div ref={mainDiv}>
         <div className="spellbook-header">
+            <input ref={fileInputRef} style={{ display: "none" }} accept=".json" type="file" onChange={event => loadJSONFile(event, setGroups)} />
             <p className="title spellbook-options">
                 Spellbook
                 <FaCirclePlus
@@ -143,12 +146,12 @@ export default function SpellBook() {
                 />
                 <FaUpload
                     style={{marginLeft: "0.5rem", cursor: "pointer", display: editing ? undefined : "none"}}
-                    onClick={() => { setGroupName(""); openModal("create-spell-group") }}
+                    onClick={() => fileInputRef.current?.click()}
                     title="Import your spellbook"
                     />
                 <FaDownload
                     style={{marginLeft: "0.5rem", cursor: "pointer", display: editing ? undefined : "none"}}
-                    onClick={() => { setGroupName(""); openModal("create-spell-group") }}
+                    onClick={() => downloadFileFromString(JSON.stringify(groups), "spellbook.json")}
                     title="Download your spellbook"
                 />
             </p>
